Score against discrete severity levels, not raw axis magnitudes
Raw, differently-scaled balance-point deviations were being summed
directly across axes in scoring.score(), letting whichever axis's
formula happens to run larger dominate every verdict regardless of real
styling significance. Contribution is now quantized into one of three
hand-picked severity levels (0/1/2, sign-preserved) before summing,
comparable across axes by construction — no external calibration data
needed. All 5 NOTES.md worked examples verified by hand to keep their
documented verdict. See docs/decisions/0010.

+++ src/fit_balance/garments.py
@@ -95,7 +95,7 @@ class AttributedReason(BaseModel):
 
     tag: str
     axis: str
-    contribution: float
+    contribution: int
     direction: Literal["+", "-"]
     item_ids: list[str]
